image_processing_service/requirements.txt
redis==5.0.8 
Pillow==10.4.0
pydantic==2.8.2
pydantic-settings==2.4.0

